Add reenter bank account number page
