+++ Android/xfers_sdk/src/main/AndroidManifest.xml
@@ -12,7 +12,8 @@
 
         <activity android:name=".view.topup.TopupBankSelectionActivity" />
         <activity android:name=".view.pay.PaymentConfirmationActivity" />
-        <activity android:name=".view.manage_banks.add_banks.AddBankAccountNumberActivity" />
+        <activity android:name=".view.manage_banks.add_banks.EnterBankAccountNumberActivity" />
+        <activity android:name=".view.manage_banks.add_banks.ReEnterBankAccountNumberActivity" />
 
         <activity
             android:name=".view.connect.ConnectPhoneActivity"
