+++ Android/xfers_sdk/src/main/res/layouts/manage_banks/layout/activity_add_bank_account_field.xml
